+++ src/hooks/index.ts
@@ -1,4 +1,1 @@
-export * from "./useCallback";
-export * from "./useDeepMemo";
-export * from "./useMemo";
-export * from "./useRef";
+export * from "./common";
